Isolate timeline editing from engine facade
edit_timeline and its composite overlay helper now live in a dedicated timeline module, leaving engine.py as a compatibility facade. The move keeps timeline track collection, trimming, merge, overlay, audio, resize, and export behavior together to avoid splitting the final coupled engine island.

Constraint: This is the highest-risk extraction; move the coupled timeline/composite code together and preserve public mcp_video.engine imports.

Rejected: Split _apply_composite_overlays into its own PR | edit_timeline is its only caller and separating them would create unnecessary coordination risk.

Confidence: medium

Scope-risk: broad

Directive: Timeline/composite overlay behavior is dense and filter-sensitive; require focused timeline tests plus broad server/e2e coverage before edits.

Tested: ruff check mcp_video/engine.py mcp_video/engine_timeline.py; ruff format --check mcp_video/engine.py mcp_video/engine_timeline.py; python3 -m pytest tests/test_engine_advanced.py::TestEditTimeline tests/test_red_team.py -k 'edit_timeline' -q --tb=short; python3 -m pytest tests/test_cli.py -k 'edit' tests/test_server.py -k 'edit' tests/test_client.py -k 'edit' -q --tb=short; python3 -m pytest tests/test_engine.py tests/test_e2e.py tests/test_server.py -q --tb=short

Not-tested: Full slow/real-media suite.
